refactor: GameResultResponseDto 생성자 수정
