+++ src/main/java/racingcar/UserRequestDto.java
@@ -1,0 +1,19 @@
+package racingcar;
+
+public final class UserRequestDto {
+
+    private String names;
+    private int count;
+
+    public UserRequestDto() {
+    }
+
+    public int getCount() {
+        return count;
+    }
+
+    public String getNames() {
+        return names;
+    }
+}
+
